conciseness

+++ examples/principal_roots.rs
@@ -1,10 +1,10 @@
 use croot::prelude::*;
 
 fn main() {
-    let x = 81;
+    let x = 81.0;
 
     // Principal 4th root of 81
-    let principal = principal_root(x as f64, 4);
+    let principal = principal_root(x, 4);
 
     println!("{principal:?}");
 }
